Move useReactorToken into gatewayAuth directory
