+++ src/pages/DataPlaneAuthReq.tsx
@@ -3,8 +3,8 @@ import { PostgrestError } from '@supabase/postgrest-js';
 import { getLiveSpecs_dataPlaneAuthReq } from 'api/liveSpecsExt';
 import { authenticatedRoutes } from 'app/routes';
 import Error from 'components/shared/Error';
+import useReactorToken from 'hooks/gatewayAuth/useReactorToken';
 import usePageTitle from 'hooks/usePageTitle';
-import useReactorToken from 'hooks/useReactorToken';
 import { useEffect, useState } from 'react';
 import { useIntl } from 'react-intl';
 import { useSearchParams } from 'react-router-dom';
